+++ stores/themeStore.ts
@@ -1,6 +1,6 @@
 import { defineStore } from 'pinia'
 
-type Theme = 'light' | 'dark';
+type Theme = 'light' | 'dark'
 
 export const useThemeStore = defineStore('theme', {
   state: () => ({
@@ -8,13 +8,15 @@ export const useThemeStore = defineStore('theme', {
   }),
   actions: {
     initTheme () {
+      // Detects system preference when loading
       const isDarkModePreferred = window.matchMedia('(prefers-color-scheme: dark)').matches
-      const themeFromLocalStorage = localStorage.getItem('theme') as Theme
+      const defaultTheme: Theme = isDarkModePreferred ? 'dark' : 'light'
 
-      this.setTheme(themeFromLocalStorage || (isDarkModePreferred ? 'dark' : 'light'))
+      // Initializes the theme with the persisted value or system preference.
+      const currentTheme = this.lightmode ? 'light' : 'dark'
+      this.setTheme(currentTheme || defaultTheme)
     },
     setTheme (newTheme: Theme) {
-      localStorage.setItem('theme', newTheme)
       document.documentElement.classList.remove('light', 'dark')
       document.documentElement.classList.add(newTheme)
       this.lightmode = newTheme === 'light'